backend/pyproject.toml
[project]
name = "local-ta-backend"
version = "0.1.0"
description = "Local TA backend with FastAPI and Ollama model management"
requires-python = ">=3.11"
authors = [{ name = "local-ta" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.5",
    "aiosqlite>=0.20",
    "chromadb>=0.5.13",
    "llama-index>=0.11",
    "llama-index-vector-stores-chroma>=0.3.0",
    "llama-index-llms-ollama>=0.3.1",
    "llama-index-embeddings-ollama>=0.3.0",
    "pypdf>=5.0.0",
    "pytesseract>=0.3.13",
    "camelot-py[cv]>=0.11.0",
    "opencv-python>=4.10.0.84",
    "python-multipart>=0.0.9",
]

[project.scripts]
ensure-ollama = "app.ollama_manager:cli_main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uv]
index-url = "https://pypi.org/simple"
package = true


